simplify response
